Added limit to groups having less than one active tag (fixes #48)

+++ flowback/group/services.py
@@ -271,6 +271,11 @@ def group_tag_update(*, user: int, group: int, tag: int, data) -> GroupTags:
     tag = get_object(GroupTags, group_id=group, id=tag)
     non_side_effect_fields = ['active']
 
+    if (GroupTags.objects.filter(group_id=group, active=True).count() <= 1
+            and tag.is_active
+            and data.get('active')):
+        raise ValidationError('Group must have at least one active tag available for users')
+
     group_tag, has_updated = model_update(instance=tag,
                                           fields=non_side_effect_fields,
                                           data=data)
@@ -281,6 +286,10 @@ def group_tag_update(*, user: int, group: int, tag: int, data) -> GroupTags:
 def group_tag_delete(*, user: int, group: int, tag: int) -> None:
     group_user_permissions(group=group, user=user, permissions=['admin'])
     tag = get_object(GroupTags, group_id=group, id=tag)
+
+    if GroupTags.objects.filter(group_id=group, active=True).count() <= 1 and tag.is_active:
+        raise ValidationError('Group must have at least one active tag available for users')
+
     tag.delete()
